Trying to extract stuff into storage
