+++ test/core/TestTsdbQuery.java
@@ -29,6 +29,8 @@ import java.util.Map;
 
 import net.opentsdb.meta.Annotation;
 import net.opentsdb.storage.MockBase;
+import net.opentsdb.storage.Span;
+import net.opentsdb.storage.SpanGroup;
 import net.opentsdb.uid.NoSuchUniqueId;
 import net.opentsdb.uid.NoSuchUniqueName;
 import net.opentsdb.uid.UniqueId;
