+++ test/storage/TestSpan.java
@@ -10,7 +10,7 @@
 // General Public License for more details.  You should have received a copy
 // of the GNU Lesser General Public License along with this program.  If not,
 // see <http://www.gnu.org/licenses/>.
-package net.opentsdb.core;
+package net.opentsdb.storage;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -19,7 +19,7 @@ import static org.powermock.api.mockito.PowerMockito.mock;
 
 import java.util.List;
 
-import net.opentsdb.storage.MockBase;
+import net.opentsdb.core.*;
 import net.opentsdb.uid.UniqueId;
 import net.opentsdb.utils.Config;
 
@@ -42,7 +42,7 @@ import com.stumbleupon.async.Deferred;
 @PowerMockIgnore({"javax.management.*", "javax.xml.*",
              "ch.qos.*", "org.slf4j.*",
              "com.sum.*", "org.xml.*"})
-@PrepareForTest({ RowSeq.class, TSDB.class, UniqueId.class, KeyValue.class, 
+@PrepareForTest({ RowSeq.class, TSDB.class, UniqueId.class, KeyValue.class,
 Config.class, RowKey.class })
 public final class TestSpan {
   private TSDB tsdb = mock(TSDB.class);
